Moved ShouldThrow and ShouldNotThrow extensions into separate classes
