+++ src/Shouldly/ShouldStaticClasses/ShouldThrowExtensions.cs
@@ -8,19 +8,6 @@ namespace Shouldly
     [ShouldlyMethods]
     public static partial class Should
     {
-        /*** ShouldThrow(Action) ***/
-        public static TException ShouldThrow<TException>(this Action actual) where TException : Exception
-        {
-            return Throw<TException>(actual);
-        }
-        public static TException ShouldThrow<TException>(this Action actual, string customMessage) where TException : Exception
-        {
-            return Throw<TException>(actual, customMessage);
-        }
-        public static TException ShouldThrow<TException>(this Action actual, [InstantHandle] Func<string> customMessage) where TException : Exception
-        {
-            return Throw<TException>(actual, customMessage);
-        }
         /*** Should.Throw(Action) ***/
         public static TException Throw<TException>([InstantHandle] Action actual) where TException : Exception
         {
@@ -48,20 +35,6 @@ namespace Shouldly
             throw new ShouldAssertException(new ExpectedShouldlyMessage(typeof(TException), customMessage).ToString());
         }
 
-        /*** ShouldNotThrow(Action) ***/
-        public static void ShouldNotThrow(this Action action)
-        {
-            NotThrow(action);
-        }
-        public static void ShouldNotThrow(this Action action, string customMessage)
-        {
-            NotThrow(action, customMessage);
-        }
-        public static void ShouldNotThrow(this Action action, [InstantHandle] Func<string> customMessage)
-        {
-            NotThrow(action, customMessage);
-        }
-
         /*** Should.NotThrow(Action) ***/
         public static void NotThrow([InstantHandle] Action action)
         {
@@ -81,20 +54,6 @@ namespace Shouldly
             {
                 throw new ShouldAssertException(new ExpectedShouldlyMessage(ex.GetType(), customMessage).ToString());
             }
-        }
-
-        /*** ShouldNotThrow(Func<T>) ***/
-        public static T ShouldNotThrow<T>(this Func<T> action)
-        {
-            return NotThrow(action);
-        }
-        public static T ShouldNotThrow<T>(this Func<T> action, string customMessage)
-        {
-            return NotThrow(action, customMessage);
-        }
-        public static T ShouldNotThrow<T>(this Func<T> action, [InstantHandle] Func<string> customMessage)
-        {
-            return NotThrow(action, customMessage);
         }
 
         /*** Should.NotThrow(Func<T>) ***/
